Change "disconnect-site" message to "site-connection-lost"
We want to be able to distinguish between a site intentionally *leaving*
a portal and a site unintentionally losing its connection with the
portal. In both cases, the site has "disconnected" from the portal, so
let's use different terminology for leaving a portal (via the
"site-did-leave" message added in b625595) and a site losing its
connection to the portal (via a "site-connection-lost" message
implemented in this commit).