Cache ICE server URLs for a *subset* of their valid lifetime
When returning ICE server URLs to a client, give the client a TTL that
is a subset of the TTL that we got from Twilio. When we get close to the
expiration time for the ICE server URLs, we want to ensure that the
client fetches new (fresh) ICE server URLs. This should ensure that a
client doesn't attempt to use ICE server URLs right at the moment that
those URLs are expiring.