Rename `username` to `login` in `IdentityProvider`
real-time-client serializes the `login` property when broadcasting peer
identities in the host. Hence, portal guests were receiving an incorrect
version of the identity which contained an empty `login` string.

Tests were not catching this because we mock all the interactions with
the identity provider, and the fake provider uses the correct `login`
property.