CORS support - allowing use directly from browser
This enables CORS headers on all requests, so teletype-client can be
used directly from within a web-browser, without any intermediate
services.

Security considerations: By enabling other domains to access the API
from the browser, it opens up the potential for a malicious site to
secretly perform tasks on behalf of the user. However, this risk is
mitigated because all requests that have side-effects or expose user
specific information also require github-oauth-token. So in order
for a site to access the API, it requires the end-user's token, just
like Atom.