Add flag to disable bugsnag on the test server.
When running tests from the teletype package we instantiate a test
server to assert the end to end integration between client and server.

This is problematic on Electron v3 because of a crash that happens due
to the bugsnag module. Since this feature is not necessary in tests we
decided to allow disabling it to work around the problem.

Refs: https://github.com/atom/teletype/issues/438
Co-Authored-By: Antonio Scandurra <as-cii@github.com>